[FIXED JENKINS-25084] Chuck no longer since update
IIUC, this regression is certainly due to @tfenelly's work
https://groups.google.com/d/msg/jenkinsci-dev/zDaX4yiWLLw/FFeLVg-2_ygJ
and
https://github.com/jenkinsci/jenkins/pull/1289/files#diff-d00c6bc3369be9674b400c1a8163acb6L209
more specifically.

As this table was no longer present, it made the next instructions
unreachable hence making the plugin not enabling the image.

Doesn't seem like the removal of the background is required to make it
work. I suppose this has become unnecessary at some point (or was it
ever?).